% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmi}
\alias{nmi}
\title{Résumé de l'échantillonnage par unité stratigraphique.}
\format{
Un \code{\link[=data.frame]{tableau de données}} de 9 lignes
 (contextes) et 7 colonnes (variables) :
 \describe{
   \item{materiau}{Nature du matériau.}
   \item{ensemble}{Ensemble chronologie.}
   \item{chronologie}{Chronologie calendaire.}
   \item{US}{Unité stratigraphique.}
   \item{NR}{Nombre de restes.}
   \item{NMI}{Nombre minimum d'individus.}
   \item{N}{Effectif de l'échantillonnage.}
 }
 Les cellules marquée d'une asterisque sont en cours d'étude.
}
\source{
Frerebeau, N. and Sacilotto, C. (2017). On Some Iberian Unfired
 Pottery Sherds from the Late Iron Age (Second Century BC).
 In Gorgues, A., Rebay-Salisbury, K. and Salisbury, R. B. (eds.),
 \emph{Material Chains in Late Prehistoric Europe and the Mediterranean -
 Time, Space and Technologies of Production}. Bordeaux: Ausonius Éditions,
 p. 157-169. Mémoires, 48.

 Sacilotto, C., 2011. \emph{De l'argile à la terre cuite, le traitement des
 productions sur un atelier de poteries ibériques à la fin de l'Age du Fer.
 Étude de l'atelier de potiers du Mas de Moreno (Foz-Calanda, Teruel,
 Espagne) des IIe et Ier siècles a.C.} Mémoire de master, Université Bordeaux
 3 Michel de Montaigne, Pessac.
}
\usage{
nmi
}
\description{
Résumé de l'échantillonnage par unité stratigraphique.
}
\seealso{
Other datasets: 
\code{\link{corpus}}
}
\concept{datasets}
\keyword{datasets}
