% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaphoCeram-package.R
\docType{package}
\name{TaphoCeram-package}
\alias{TaphoCeram-package}
\alias{_PACKAGE}
\alias{TaphoCeram}
\title{TaphoCeram: Replication of Frerebeau, N., Ben Amara, A. and Cantin, N. (2020)}
\description{
Research compendium of Frerebeau, N., Ben Amara, A. and Cantin, N.
    (2020) <doi:10.4000/archeosciences.7220> [in french]. The compendium
    contains all data, code, and text associated with this paper.
}
\details{
\tabular{ll}{
 \strong{Package:} \tab TaphoCeram \cr
 \strong{Type:} \tab Compendium \cr
 \strong{Version:} \tab 1.3 \cr
 \strong{License:} \tab GPL-3 \cr
 \strong{DOI:} \tab \href{https://doi.org/10.5281/zenodo.2613512}{10.5281/zenodo.2613512}
}
}
\note{
This work received a state financial support managed by the Agence Nationale
de la Recherche (France) throught the program \emph{Investissements d'avenir}
(ref. ANR-10-LABX-52).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nfrerebeau/TaphoCeram}
  \item Report bugs at \url{https://github.com/nfrerebeau/TaphoCeram/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
 Nicolas Frerebeau \tab \emph{UMR 5060 IRAMAT-CRP2A, Université Bordeaux Montaigne, France} \cr
 Ayed Ben Amara \tab \emph{UMR 5060 IRAMAT-CRP2A, Université Bordeaux Montaigne, France} \cr
 Nadia Cantin \tab \emph{UMR 5060 IRAMAT-CRP2A, Université Bordeaux Montaigne, France}
}

\strong{Corresponding author}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

UMR 5060 IRAMAT-CRP2A\cr
Maison de l'Archéologie\cr
Universite Bordeaux Montaigne\cr
33607 Pessac cedex, France
}
\keyword{internal}
