% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/visualize-PCA.R
\name{visualize}
\alias{visualize}
\alias{visualize.PCA}
\title{Plot PCA results}
\usage{
visualize(x, ...)

\method{visualize}{PCA}(x, axes = c(1, 2), map = c("individuals",
  "variables", "eigenvalues"), extra = NULL, select = NULL,
  group = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link[FactoMineR]{PCA}} or
\code{\link[FactoMineR]{CA}}.}

\item{...}{Further arguments passed to other methods.}

\item{axes}{A length-two \code{\link{numeric}} vector specifying the
components to plot.}

\item{map}{A \code{\link{character}} string or vector of strings specifying
the graph to plot (see below). Any unambiguous substring can be given.}

\item{extra}{A string or vector of \code{\link{character}} strings
specifiying the supplementary data to plot (see below).
Any unambiguous substring can be given.
If \code{NULL} (default) no extra information is plotted.}

\item{select}{A \code{\link{numeric}} or \code{\link{character}} vector
specifying a part of the elements that are drawn (see details in
\code{\link[FactoMineR]{PCA}} and \code{\link[FactoMineR]{CA}}).
If \code{NULL} (default) no selection is made.}

\item{group}{A \code{\link{character}} vector of categories or a length-one
\code{\link{numeric}} vector specifiying the supplementary categorial
variable from which to color the individuals.
The elements are coerced to characters by \code{\link{as.character}}.}
}
\description{
Visualize multivariate exploratory data analysis results.
}
\section{PCA}{

 \code{map} should be one of the following:
 \describe{
  \item{individuals}{Plots individuals factor map along the specified
   \code{axes}.}
  \item{variables}{Plots variables factor map along the specified
   \code{axes}.}
  \item{eigenvalues}{Plots eigenvalues and the cumulative percentage of
   variance.}
 }
 \code{extra} should be one or more of the following:
 \describe{
  \item{individuals}{Add supplementary individuals.}
  \item{qualitative}{Add supplementary categories.}
  \item{quantitative}{Add supplementary continuous variables.}
 }
}

\section{CA}{

 \code{map} should be one or more of the following:
 \describe{
  \item{rows}{Plots rows along the specified \code{axes}.}
  \item{columns}{Plots columns along the specified \code{axes}.}
 }
 \code{extra} should be one or more of the following:
 \describe{
  \item{rows}{Add supplementary rows.}
  \item{columns}{Add supplementary columns.}
  \item{qualitative}{Add supplementary categories.}
  \item{quantitative}{Add supplementary continuous variables.}
 }
}

\seealso{
\link[FactoMineR]{PCA}, \link[FactoMineR]{CA}
}
\author{
N. Frerebeau
}
