% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archeosciences2019-package.R
\docType{package}
\name{archeoscience2019-package}
\alias{archeoscience2019-package}
\alias{_PACKAGE}
\alias{archeoscience2019}
\title{archeosciences2019: Replication of Frerebeau, N., Ben Amara, A. and Cantin, N. (2019)}
\description{
Research compendium of Frerebeau, N., Ben Amara, A. and Cantin, N. 
    (2019) <doi:xxx> [in french]. The compendium contains all data, code, and 
    text associated with this paper.
}
\details{
\tabular{ll}{
 \strong{Package:} \tab archeosciences2019 \cr
 \strong{Type:} \tab Compendium \cr
 \strong{Version:} \tab 1.1 \cr
 \strong{License:} \tab GPL-3 \cr
 \strong{DOI:} \tab \href{https://doi.org/10.5281/zenodo.2613512}{10.5281/zenodo.2613512}
}
}
\note{
This work received a state financial support managed by the Agence Nationale
de la Recherche (France) throught the program \emph{Investissements d'avenir}
(ref. ANR-10-LABX-52).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/nfrerebeau/archeosciences2019}
  \item Report bugs at \url{http://github.com/nfrerebeau/archeosciences2019/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
 Nicolas Frerebeau \tab \emph{UMR 5060 IRAMAT-CRP2A, Université Bordeaux Montaigne, France} \cr
 Ayed Ben Amara \tab \emph{UMR 5060 IRAMAT-CRP2A, Université Bordeaux Montaigne, France} \cr
 Nadia Cantin \tab \emph{UMR 5060 IRAMAT-CRP2A, Université Bordeaux Montaigne, France}
}

\strong{Corresponding author}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

UMR 5060 IRAMAT-CRP2A\cr
Maison de l'Archéologie\cr
Universite Bordeaux Montaigne\cr
33607 Pessac cedex, France
}
\keyword{internal}
